Move the 3 PlexusTestCase classes into a new test support module
