+++ pom.xml
@@ -80,6 +80,7 @@ under the License.
   </properties>
 
   <modules>
+    <module>maven-test-support</module>
     <module>maven-plugin-api</module>
     <module>maven-builder-support</module>
     <module>maven-model</module>
@@ -181,6 +182,11 @@ under the License.
     <dependencies>
       <!--  Maven Modules -->
       <!--bootstrap-start-comment-->
+      <dependency>
+        <groupId>org.apache.maven</groupId>
+        <artifactId>maven-test-support</artifactId>
+        <version>${project.version}</version>
+      </dependency>
       <dependency>
         <groupId>org.apache.maven</groupId>
         <artifactId>maven-model</artifactId>
